+++ python/packages/core/agent_framework/openai/_responses_client.py
@@ -455,13 +455,18 @@ class OpenAIBaseResponsesClient(OpenAIBase, BaseChatClient):
                         },
                     }
                 if content.has_top_level_media_type("application"):
-                    return {
+                    filename = getattr(content, "filename", None) or (
+                        content.additional_properties.get("filename")
+                        if hasattr(content, "additional_properties") and content.additional_properties
+                        else None
+                    )
+                    file_obj = {
                         "type": "input_file",
-                        "filename": content.additional_properties["filename"]
-                        if content.additional_properties and "filename" in content.additional_properties
-                        else "uploaded_file",
                         "file_data": content.uri,
                     }
+                    if filename:
+                        file_obj["filename"] = filename
+                    return file_obj
                 return {}
             case FunctionCallContent():
                 return {
