+++ python/samples/getting_started/multimodal_input/azure_chat_multimodal.py
@@ -1,7 +1,6 @@
 # Copyright (c) Microsoft. All rights reserved.
 
 import asyncio
-import base64
 
 from agent_framework import ChatMessage, DataContent, Role, TextContent
 from agent_framework.azure import AzureOpenAIChatClient
@@ -9,7 +8,7 @@ from azure.identity import AzureCliCredential
 
 
 async def test_image() -> None:
-    """Test image analysis with Azure."""
+    """Test image analysis with Azure OpenAI."""
     # For authentication, run `az login` command in terminal or replace AzureCliCredential with preferred
     # authentication option. Requires AZURE_OPENAI_ENDPOINT and AZURE_OPENAI_CHAT_DEPLOYMENT_NAME
     # environment variables to be set.
@@ -27,31 +26,11 @@ async def test_image() -> None:
     print(f"Image Response: {response}")
 
 
-async def test_pdf() -> None:
-    """Test PDF document analysis with Azure."""
-    client = AzureOpenAIChatClient(credential=AzureCliCredential())
-
-    pdf_uri = create_sample_pdf()
-    message = ChatMessage(
-        role=Role.USER,
-        contents=[
-            TextContent(text="What information can you extract from this document?"),
-            DataContent(
-                uri=pdf_uri,
-                media_type="application/pdf",
-                additional_properties={"filename": "employee_report.pdf"}
-            ),
-        ],
-    )
-
-    response = await client.get_response(message)
-    print(f"PDF Response: {response}")
-
-
 async def main() -> None:
-    print("=== Testing Azure Multimodal ===")
+    print("=== Testing Azure OpenAI Multimodal ===")
+    print("Testing image analysis (supported by Chat Completions API)")
+    print("Note: For PDF support, use the Azure OpenAI Responses API instead.")
     await test_image()
-    await test_pdf()
 
 
 def create_sample_image() -> str:
@@ -59,79 +38,6 @@ def create_sample_image() -> str:
     # This is a tiny red pixel in PNG format
     png_data = "iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAADUlEQVR42mP8/5+hHgAHggJ/PchI7wAAAABJRU5ErkJggg=="
     return f"data:image/png;base64,{png_data}"
-
-
-def create_sample_pdf() -> str:
-    """Create a minimal PDF document for testing."""
-    pdf_content = """%PDF-1.4
-1 0 obj
-<<
-/Type /Catalog
-/Pages 2 0 R
->>
-endobj
-
-2 0 obj
-<<
-/Type /Pages
-/Kids [3 0 R]
-/Count 1
->>
-endobj
-
-3 0 obj
-<<
-/Type /Page
-/Parent 2 0 R
-/MediaBox [0 0 612 792]
-/Resources <<
-/Font <<
-/F1 4 0 R
->>
->>
-/Contents 5 0 R
->>
-endobj
-
-4 0 obj
-<<
-/Type /Font
-/Subtype /Type1
-/BaseFont /Helvetica
->>
-endobj
-
-5 0 obj
-<<
-/Length 44
->>
-stream
-BT
-/F1 12 Tf
-100 700 Td
-(Employee Review: John Smith) Tj
-ET
-endstream
-endobj
-
-xref
-0 6
-0000000000 65535 f
-0000000009 00000 n
-0000000058 00000 n
-0000000115 00000 n
-0000000274 00000 n
-0000000361 00000 n
-trailer
-<<
-/Size 6
-/Root 1 0 R
->>
-startxref
-456
-%%EOF"""
-    pdf_b64 = base64.b64encode(pdf_content.encode()).decode()
-    return f"data:application/pdf;base64,{pdf_b64}"
 
 
 if __name__ == "__main__":
